+++ app/actions/user_action.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+class UserAction
+
+  def perform
+    # write the logic here
+  end
+
+end
