+++ app/schemas/user_schema.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+class UserSchema
+
+    attr_writer :name, :email
+
+end
