+++ lib/generators/business_logic/action/test_unit/test_unit_generator.rb
@@ -1,9 +1,13 @@
 # frozen_string_literal: true
 
+require_relative "../../../parser"
+
 module BusinessLogic
   module Generators
     module Action
       class TestUnitGenerator < Rails::Generators::NamedBase
+
+        include Parser
 
         source_root File.expand_path("templates", __dir__)
