+++ lib/generators/business_logic/schema/templates/schema.rb.erb.tt
@@ -3,7 +3,7 @@
 class <%= class_name %>Schema
 
   <%- if @parsed_fields.any? -%>
-    attr_writer <%= @parsed_fields.map(&:name).join(", ") %>
+    attr_writer <%= @parsed_fields.map { |field| ":#{field.name}" }.join(", ") %>
   <%- end -%>
 
 end
